# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SevaSetu API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ServicePriority =
  (typeof ServicePriority)[keyof typeof ServicePriority];

export const ServicePriority = {
  P0: "P0",
  P1: "P1",
  P2: "P2",
} as const;

export interface Service {
  id: number;
  category: string;
  name: string;
  description: string;
  avgMinPrice: number;
  avgMaxPrice: number;
  priority: ServicePriority;
  iconName: string;
}

export type BookingTimeSlot =
  (typeof BookingTimeSlot)[keyof typeof BookingTimeSlot];

export const BookingTimeSlot = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  customerName: string;
  customerPhone: string;
  serviceCategory: string;
  serviceName: string;
  address: string;
  scheduledDate: string;
  timeSlot: BookingTimeSlot;
  description: string;
  status: BookingStatus;
  craftsmanId?: number | null;
  craftsmanName?: string | null;
  totalAmount?: number | null;
  platformFee?: number | null;
  convenienceFee?: number | null;
  rating?: number | null;
  review?: string | null;
  createdAt: string;
}

export type CreateBookingBodyTimeSlot =
  (typeof CreateBookingBodyTimeSlot)[keyof typeof CreateBookingBodyTimeSlot];

export const CreateBookingBodyTimeSlot = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;

export interface CreateBookingBody {
  customerName: string;
  customerPhone: string;
  serviceCategory: string;
  serviceName: string;
  address: string;
  scheduledDate: string;
  timeSlot: CreateBookingBodyTimeSlot;
  description: string;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
  craftsmanId?: number | null;
  totalAmount?: number;
  platformFee?: number;
  convenienceFee?: number;
  rating?: number;
  review?: string;
}

export interface Craftsman {
  id: number;
  name: string;
  phone: string;
  photoUrl?: string | null;
  skills: string[];
  serviceAreas: string[];
  rating: number;
  totalJobs: number;
  isVerified: boolean;
  isAvailable: boolean;
  experience: number;
  bio?: string | null;
  joinedAt: string;
}

export interface CreateCraftsmanBody {
  name: string;
  phone: string;
  photoUrl?: string;
  skills: string[];
  serviceAreas: string[];
  experience: number;
  bio?: string;
}

export interface UpdateCraftsmanBody {
  name?: string;
  phone?: string;
  photoUrl?: string;
  skills?: string[];
  serviceAreas?: string[];
  isVerified?: boolean;
  isAvailable?: boolean;
  experience?: number;
  bio?: string;
}

export type DashboardSummaryBookingsByCategoryItem = {
  category: string;
  count: number;
};

export interface DashboardSummary {
  totalBookings: number;
  pendingBookings: number;
  completedBookings: number;
  activeBookings: number;
  totalCraftsmen: number;
  verifiedCraftsmen: number;
  availableCraftsmen: number;
  gmvThisMonth: number;
  revenueThisMonth: number;
  avgRating: number;
  bookingsThisWeek: number;
  bookingsByCategory: DashboardSummaryBookingsByCategoryItem[];
}

export interface RevenueDay {
  date: string;
  gmv: number;
  revenue: number;
  completedJobs: number;
}

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  limit?: number;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListCraftsmenParams = {
  skill?: string;
  available?: boolean;
};

export type GetRevenueReportParams = {
  days?: number;
};